fix code smells detected by sonar

+++ utils.h
@@ -178,7 +178,7 @@ extern size_t strlcat(char *dst, const char *src, size_t siz);
 extern char *trimr(char * const buf);
 extern char *lowercase(char * const str);
 extern char *uppercase(char * const str);
-extern int unicode(char **dst, const char * const src);
+extern size_t unicode(char **dst, const char * const src);
 extern void *zmalloc(size_t size) __attribute__((warn_unused_result, malloc, alloc_size(1)));
 extern char *urlencode(const char * const str) __attribute__((warn_unused_result));
